pyproject.toml
[project]
name = "oncolearn"
version = "0.1.0"
description = "A comprehensive toolkit for cancer genomics analysis and biomarker discovery"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "hdbscan>=0.8.41",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "plotly>=6.5.1",
    "pyreadr>=0.5.4",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
    "llvmlite>=0.40.0,<0.46.0",
    "transformers>=4.57.3",
    "pillow>=12.1.0",
    "diffusers>=0.36.0",
    "pydicom>=3.0.1",
    "simpleitk>=2.5.3",
    "scipy>=1.16.3",
    "pyyaml>=6.0.3",
]

[project.scripts]
download = "scripts.data.download:main"

[dependency-groups]
dev = [
    "oncolearn",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "autopep8>=2.3.2",
    "ruff>=0.14.11",
]

[project.optional-dependencies]
cpu = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
cu128 = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
cu130 = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
]
rocm = [
  "torch>=2.9.1",
  "torchvision>=0.24.1",
  "pytorch-triton-rocm"
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
    { extra = "cu130" },
    { extra = "rocm" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
  { index = "pytorch-cu130", extra = "cu130" },
  { index = "pytorch-rocm", extra = "rocm" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
  { index = "pytorch-cu130", extra = "cu130" },
  { index = "pytorch-rocm", extra = "rocm" },
]
pytorch-triton-rocm = [
  { index = "pytorch-rocm", extra = "rocm" },
]
oncolearn = { workspace = true, editable = true }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[[tool.uv.index]]
name = "pytorch-rocm"
url = "https://download.pytorch.org/whl/rocm6.4"
explicit = true

[tool.uv.workspace]
members = [
    ".",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
